% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_ka.R
\name{calculate_ka_statistical_moments}
\alias{calculate_ka_statistical_moments}
\title{Calculate absorption rate constant (ka) using statistical moments}
\usage{
calculate_ka_statistical_moments(x, y, nlastpoints)
}
\arguments{
\item{x}{A numeric vector representing the time points.}

\item{y}{A numeric vector representing the concentrations corresponding to the time points.}

\item{nlastpoints}{An integer specifying the number of last points to be used for the slope calculation of the terminal elimination rate constant.}
}
\value{
A numeric value representing the estimated absorption rate constant (ka).
}
\description{
Calculates the absorption rate constant (ka) for oral drug administration using statistical moments.
It calculates the area under the concentration-time curve (AUC) and the area under the moment curve (AUMC) based on non-compartmental analysis (NCA).
The absorption rate constant is derived from the mean absorption time (MAT), which is calculated using the difference between
the mean residence time (MRT) for oral and intravenous (IV) administration.
}
\details{
The function performs the following steps:
\itemize{
\item It uses the \code{nca.iv.normalised} function to calculate the area under the curve (AUC), the area under the moment curve (AUMC), and the mean residence time (MRT) for IV administration.
\item The MRT for oral administration is calculated as the ratio of AUMC to AUC.
\item The mean absorption time (MAT) is calculated as the difference between the MRT for oral and IV administration.
\item Finally, the absorption rate constant (ka) is estimated as the inverse of MAT.
}
}
\examples{
# Example data from Oral_1CPT dataset (first subject and single dose)
dat <- Oral_1CPT[Oral_1CPT$ID == 1 & Oral_1CPT$SD == 1 & Oral_1CPT$EVID == 0, ]
dat <- data.frame(TIME = dat$TIME, DV = dat$DV)

# Calculate ka using statistical moments with the last 4 points for lambda_z
calculate_ka_statistical_moments(x = dat$TIME, y = dat$DV, nlastpoints = 4)

}
