% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initsControl.R
\name{fallback_control}
\alias{fallback_control}
\title{Control settings for fallback rules in parameter estimation}
\usage{
fallback_control(
  enable_ka_fallback = TRUE,
  sigma_method_additive = "model",
  sigma_method_proportional = "model",
  sigma_fallback_fraction = 0.2
)
}
\arguments{
\item{enable_ka_fallback}{Logical. Whether to fallback to ka = 1 if invalid.}

\item{sigma_method_additive}{Method for additive sigma ("model" or "fixed_fraction").}

\item{sigma_method_proportional}{Method for proportional sigma ("model" or "fixed_fraction").}

\item{sigma_fallback_fraction}{Numeric. e.g. 0.2 (20\% of DV mean).}
}
\value{
A list of fallback control parameters.
}
\description{
Control settings for fallback rules in parameter estimation
}
\examples{
fallback_control(sigma_method_additive = "fixed_fraction", sigma_fallback_fraction = 0.25)
}
