% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_life_estimated.R
\name{half_life_estimated}
\alias{half_life_estimated}
\title{Estimate half-life from given pharmacokinetic Data}
\usage{
half_life_estimated(dat, nlastpoints = 3, nbins = 10, route = "bolus")
}
\arguments{
\item{nlastpoints}{An integer specifying the number of last data points to be used for
linear regression in the estimation.}

\item{fdat}{A data frame containing the pharmacokinetic data.}
}
\value{
A named vector with the estimated half-life values from each data subset, including:
\itemize{
\item \code{half_life_median}: The median of all positive half-life estimates from the subsets.
\item \code{half_life_fd}: The half-life estimate from first dose data, if available.
\item \code{half_life_md}: The half-life estimate from repeated dose data, if available.
\item \code{half_life_all}: The half-life estimate from the complete dataset.
}
}
\description{
Estimates the half-life based on pharmacokinetic data by linear regression on terminal phase,
considering different scenarios: data after the first dose, data after repeated doses,
and data including both first and repeated doses.
}
\details{
The function performs the following steps:
\enumerate{
\item Splits the input data into subsets based on dosing intervals.
\item Applies time binning to normalize the concentration-time data in each subset.
\item Performs linear regression on the log-transformed concentration of the last \code{nlastpoints}
to calculate the elimination rate constant (\code{ke}) and the half-life (\code{half_life}).
\item Estimates the half-life separately for:
\itemize{
\item First dose interval data
\item Repeated dose interval data
\item All data combined
}
\item Returns the median of the positive half-life estimates as \code{half_life_median}.
}
}
\examples{
# Example usage:
dat <- Bolus_1CPT
dat<- processData(dat)$dat
half_life_estimated(fdat)

}
