% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_ka.R
\name{ka_calculation_sd}
\alias{ka_calculation_sd}
\title{Calculate absorption rate constant (ka) in a single-dose one-compartment model}
\usage{
ka_calculation_sd(cl, ke, t, Ct, Fbio = 1, Dose)
}
\arguments{
\item{cl}{Numeric. Clearance of the drug (in L/hr).}

\item{ke}{Numeric. Elimination rate constant (in 1/hr).}

\item{t}{Numeric. Time after drug administration (in hours) at which the concentration is measured.}

\item{Ct}{Numeric. Observed concentration of the drug at time \code{t} (in mg/L).}

\item{Fbio}{Numeric. Bioavailability fraction (default = 1, meaning 100\% bioavailability).}

\item{Dose}{Numeric. Administered dose of the drug (in mg).}
}
\value{
A list containing the following components:
\item{ka}{The calculated absorption rate constant.}
\item{full_solution}{The full solution object returned by the \code{uniroot()} function, which includes additional details about the root-finding process.}
\item{message}{A character string providing status information about the estimation, including warnings if the observed concentration exceeds model-predicted limits or if numerical errors occurred.}
}
\description{
Calculates the absorption rate constant (\code{ka}) for a drug administered orally
using a one-compartment model with first-order absorption and elimination kinetics.
The calculation is based on known values of clearance (\code{cl}), elimination rate constant
(\code{ke}), time (\code{t}), observed concentration (\code{Ct}), bioavailability (\code{Fbio}),
and dose (\code{Dose}).
}
\details{
This function uses a one-compartment model with first-order absorption and elimination to
estimate the absorption rate constant (\code{ka}). The function solves the following equation
for \code{ka} numerically using \code{uniroot()}:

\deqn{Ct = \frac{Fbio \cdot Dose \cdot ka}{Vd \cdot (ka - ke)} \left( e^{-ke \cdot t} - e^{-ka \cdot t} \right)}

The \code{uniroot()} function is used to find the value of \code{ka} that makes the difference
between the predicted and observed concentrations equal to zero. The reasonable range for
\code{ka} is set as \link{0.01, 1000} for root-finding.

Additional safeguards include ensuring that \code{ka > ke} to avoid flip-flop kinetics, and rejecting
values of \code{Ct} that exceed the model-predicted maximum concentration. In such cases, the function
returns \code{NA} for \code{ka} and a descriptive \code{message}.
}
\examples{
# Example usage:
# Data from first point in simulated dataset Oral_1CPT
ka_result <- ka_calculation_sd(cl = 3.62, ke = 0.0556, t = 0.5, Ct = 310, Dose = 60000)
print(ka_result$ka)

}
