% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_ka.R
\name{ka_wanger_nelson}
\alias{ka_wanger_nelson}
\title{Calculate the Absorption Rate Constant (ka) using the Wanger-Nelson Method}
\usage{
ka_wanger_nelson(dat, nlastpoints, nca.out)
}
\arguments{
\item{dat}{A data frame with at least two columns: \code{DV} (drug concentration)
and \code{TIME} (time points of measurement).}

\item{nlastpoints}{An integer specifying the number of last points to be used for slope calculation
of the elimination rate constant (ke). However, this argument is not used in the current function logic but remains for future enhancements.}
}
\value{
A numeric value representing the estimated absorption rate constant (ka).
}
\description{
Computes the absorption rate constant (ka) using the Wanger-Nelson method.
The function makes use of non-compartmental analysis (NCA) function part to calculate the area under the curve (AUC) and the elimination rate constant (ke).
Linear regression on the logarithm of the remaining fraction of drug is used to estimate the absorption rate constant (ka).
}
\details{
The function calculates the absorption rate constant (ka) as follows:
\itemize{
\item The function uses the \code{nca.iv.normalised} method to calculate the area under the curve (AUC) and elimination rate constant (ke).
\item AUC is calculated using trapezoidal integration for the observed data points.
\item The fraction of drug absorbed is calculated based on ke and the cumulative AUC.
\item The logarithm of the remaining fraction of the drug is then modeled using linear regression to determine the absorption rate constant (ka).
}

This method assumes that the drug absorption follows first-order kinetics.
}
\examples{
# Example 1.  data frame with drug concentration (DV) and time (TIME)
data_example <- data.frame(
  TIME = c(0, 0.5, 1, 2, 3, 4, 5, 6, 8, 10),
  DV = c(0, 5, 10, 30, 40, 45, 30, 20, 10, 5 )
)

# Calculate ka using the Wanger-Nelson method with the last 4 data points
ka_wanger_nelson(data_example, 4)$ka

# Example 2. data frame from nlmixr2data
dat<-Oral_1CPT[Oral_1CPT$ID==1 &Oral_1CPT$SD==1&Oral_1CPT$EVID==0,]
dat<-data.frame(TIME=dat$TIME,DV=dat$DV)
ka_wanger_nelson(dat=dat,nlastpoints=4)$ka
ka_wanger_nelson(dat=dat,nlastpoints=4)$dat_out_wanger_nelson

}
