% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tad.R
\name{mark_dose_number}
\alias{mark_dose_number}
\title{Mark dose number}
\usage{
mark_dose_number(dat)
}
\arguments{
\item{dat}{PK dataset containing columns:
\itemize{
\item \strong{Required}: \code{ID}, \code{TIME}, \code{EVID}, \code{AMT}
\item \strong{Optional}: \code{resetflag} (default=1 if missing), \code{CMT}
}}
}
\value{
A modified dataframe with added \code{dose_number} column
}
\description{
Marks each dosing event in the dataset with a sequential dose number,
based on rows identified by \code{EVID}. Dose numbering is performed within
groups defined by \code{ID} and \code{resetflag}, ensuring numbering restarts
appropriately for each group.
}
\examples{
mark_dose_number(Bolus_1CPT)
mark_dose_number(Infusion_1CPT)
mark_dose_number(Oral_1CPT)

}
