% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pooled_nca.R
\name{nca_control}
\alias{nca_control}
\title{Control settings for non-compartmental analysis (NCA)}
\usage{
nca_control(
  trapezoidal.rule = c("linear_up_log_down", "linear"),
  duration = NULL,
  nlastpoints = 3,
  slope.method = "bestfitforce"
)
}
\arguments{
\item{nlastpoints}{Integer. Number of terminal points for half-life regression.}

\item{slope.method}{Method for estimating terminal slope (\eqn{\lambda_z}):
\itemize{
\item \code{"bestfitforce"} - Force estimation using decreasing number of terminal points if best-fit fails (default)
\item \code{"bestfit"} - Use automated best-fit selection based on adjusted R-squared
}}

\item{trapezoidal_rule}{Integer. 1 = linear, 2 = linear-up log-down (default = 1).}
}
\value{
A list with NCA control parameters.
}
\description{
Control settings for non-compartmental analysis (NCA)
}
\examples{
nca_control()
}
