% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk.time.binning.R
\name{pk.time.binning}
\alias{pk.time.binning}
\title{Create time bins for pooling a pharmacokinetic data}
\usage{
pk.time.binning(dat, nbins)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data with columns for time after dose (tad) and normalised concentration (DVstd).}

\item{nbins}{The number of bins to create.}
}
\value{
A list containing two data frames: \code{test.pool.normalised} with median time and concentration for each bin, and \code{test.pool.group} with the maximum time in each bin.
}
\description{
Creates time bins for the provided pharmacokinetic data and calculates the median time and concentration for each bin.
}
\examples{
dat <- Bolus_1CPT
dat <- nmpkconvert(dat)
dat <- calculate_tad(dat)
dat$DVstd<-dat$DV/dat$dose
pk.time.binning(dat,8)
}
