% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_graphcal_iv.R
\name{run_graphcal_iv}
\alias{run_graphcal_iv}
\title{Run graphical calculation for intravenous data}
\usage{
run_graphcal_iv(dat, fdobsflag, nbins, nlastpoints)
}
\arguments{
\item{dat}{A data frame containing the intravenous pharmacokinetic data.}

\item{fdobsflag}{A flag indicating whether the dataset includes first-dose observations. Set to 1 if the data includes first-dose observations, otherwise set to 0.}

\item{nbins}{The number of bins for time binning.}

\item{nlastpoints}{The number of last points to be used for slope linear regression.}
}
\value{
A data frame with calculated clearance, volume of distribution, slope, and time spent for the calculation.
}
\description{
Performs a graphical calculation of clearance and volume of distribution based on provided pharmacokinetic data with concentration normalised by dose
}
\examples{
dat <- Bolus_1CPT
dat <- nmpkconvert(dat)
dat <- calculate_tad(dat)
run_graphcal_iv(dat, fdobsflag = 1, nbins = 8, nlastpoints = 4)
}
