% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nca.normalised.R
\name{run_nca.normalised}
\alias{run_nca.normalised}
\title{Run non-compartmental analysis (NCA) for intravenous pharmacokinetic data}
\usage{
run_nca.normalised(
  dat,
  nlastpoints = 3,
  trapezoidal.rule = 1,
  nbins = 8,
  fdobsflag = 1,
  sdflag = 1,
  route = "bolus"
)
}
\arguments{
\item{dat}{A data frame containing the intravenous pharmacokinetic data which include at least columns for ID, TIME, EVID, AMT, DV, and dose_number.}

\item{nlastpoints}{The number of last points to be used for slope calculation.}

\item{trapezoidal.rule}{A numeric value indicating the method used for calculating AUC using the trapezoidal rule.
1 for the standard linear trapezoidal method.
2 for the linear up and logarithmic down method, where the concentration increase is treated linearly,
and the concentration decrease is treated logarithmically.}

\item{nbins}{The number of bins for time binning.}

\item{fdobsflag}{A flag indicating whether the dataset includes first-dose observations. Set to 1 if the data includes first-dose observations, otherwise set to 0.}
}
\value{
A list containing the results of the NCA calculation for the first dose (\code{nca.fd.results}) and for all data (\code{nca.results}).
}
\description{
Perform a non-compartmental analysis (NCA) for intravenous dosing data with normalised concentration with dose, calculating clearance, volume of distribution, slope, and half-life for all pooled data, as well as for only pooled first-dose data if available.
}
\examples{
dat <- theo_sd
dat <- processData(dat)$dat
run_nca.normalised(dat, nlastpoints = 3, trapezoidal.rule=1,nbins = 8, fdobsflag = 1,sdflag=1)
run_nca.normalised(dat, nlastpoints = 3, trapezoidal.rule=2,nbins = 8, fdobsflag = 1,sdflag=1)
}
