% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modelfit.R
\name{run_npd_1cmpt_iv}
\alias{run_npd_1cmpt_iv}
\title{Run and evaluate one-compartment IV Model}
\usage{
run_npd_1cmpt_iv(
  dat,
  est.method = "nls",
  input.cl = exp(1),
  input.vd = exp(1),
  input.add = 1
)
}
\arguments{
\item{dat}{A data frame containing intravenous pharmacokinetic data. It should include
columns such as \code{ID}, \code{EVID}, \code{DV}, \code{dose}, and \code{AMT} and other necessary variables.}

\item{est.method}{The estimation method in nlmixr2 to use (e.g., "nls", "nlm", focei"...). The default value is "nls".}

\item{input.cl}{A numeric value for the initial estimate of clearance (CL). Default is exp(1).}

\item{input.vd}{A numeric value for the initial estimate of volume of distribution (Vd). Default is exp(1).}

\item{input.add}{A numeric value for the additive error model. Default is Default is 1.}
}
\value{
A list containing the following elements:
\item{npd.1cmpt_results}{A data frame with the estimated CL and Vd, as well
as the time spent during estimation.}
\item{npd.1cmpt.APE}{The absolute prediction error (APE).}
\item{npd.1cmpt.MAE}{The mean absolute error (MAE).}
\item{npd.1cmpt.MAPE}{The mean absolute percentage error (MAPE).}
\item{npd.1cmpt.RMSE}{The root mean square error (RMSE).}
\item{npd.1cmpt.rRMSE}{The relative root mean square error (rRMSE).}
\item{nnpd.1cmpt.list}{The full output list from the \code{Fit_1cmpt_iv} function.}
}
\description{
Perform parameter estimation with naive pooled data approach analysis for a one-compartment intravenous model, and evaluate the model fitting performance by the absolute prediction error (APE) and mean absolute prediction error (MAPE).
}
\details{
This function fits the one-compartment IV model to the given dataset using
the specified estimation methods. It excludes rows where \code{EVID == 2}.
After fitting the model, the function returns the estimated CL and Vd along
with metrics to assess model fit.
}
\examples{
\dontrun{
# Example usage:
result <- run_npd_1cmpt_iv(dat = Bolus_1CPT,input.cl=4,input.vd=70)
}

}
