% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modelfit.R
\name{run_npd_1cmpt_mm_iv}
\alias{run_npd_1cmpt_mm_iv}
\title{Conduct and evaluate one-compartment with Michaelis-Menten Kinetics modelling}
\usage{
run_npd_1cmpt_mm_iv(dat, est.method, npdmm_inputcl, npdmm_inputvd)
}
\arguments{
\item{dat}{A data frame containing the intravenous pharmacokinetic data.}

\item{est.method}{The estimation method in nlmixr2 to use (e.g., "nls", "nlm", focei"...). The default value is "nls".}

\item{npdmm_inputcl}{Pre-estimated clearance for the intravenous data, used to determine the boundary values for Vmax and Km.}

\item{npdmm_inputvd}{The initial estimate of volume of distribution.}
}
\value{
A list containing the results of the model fitting, including the fitted parameters (\code{npd.1cmpt.mm_results} and \code{npd.1cmpt.mm_results.2}), APE (\code{npd.1cmpt.mm.APE} and \code{npd.1cmpt.mm.APE.2}), MAPE (\code{npd.1cmpt.mm.MAPE} and \code{npd.1cmpt.mm.MAPE.2}), and the model fitting lists (\code{npd.1cmpt.mm.list} and \code{npd.1cmpt.mm.list2}).
}
\description{
Perform parameter estimation with naive pooled data approach analysis for a one-compartment intravenous model, and use provided or calculated Vmax, Km and volume of distribution estimates to fit the model and calculate the absolute prediction error (APE) and mean absolute prediction error (MAPE).
}
\examples{
dat <- Bolus_1CPTMM
run_npd_1cmpt_mm_iv(dat=dat,est.method="nls",input.cl=4,input.vd=70)
}
