% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modelfit.R
\name{run_npd_2cmpt_oral}
\alias{run_npd_2cmpt_oral}
\title{Run and evaluate two-compartment oral Model}
\usage{
run_npd_2cmpt_oral(
  dat,
  est.method = "nls",
  input.ka = 1,
  input.cl = 1,
  input.vc2cmpt = 1,
  input.vp2cmpt = 1,
  input.q2cmpt = 1,
  input.add = 1
)
}
\arguments{
\item{dat}{A data frame containing intravenous pharmacokinetic data. It should include
columns such as \code{ID}, \code{EVID}, \code{DV}, \code{dose}, and \code{AMT} and other necessary variables.}

\item{est.method}{The estimation method in nlmixr2 to use (e.g., "nls", "nlm", focei"...). The default value is "nls".}

\item{input.ka}{A numeric value for the initial estimate of absorption rate constant. Default is 1.}

\item{input.cl}{A numeric value for the initial estimate of clearance (CL). Default is 1.}

\item{input.vc2cmpt}{A numeric value for the initial estimate of the volume of distribution in the central compartment (Vc). Default is 1.}

\item{input.vp2cmpt}{A numeric value for the initial estimate of the volume of distribution in the peripheral compartment (Vp). Default is 1.}

\item{input.q2cmpt}{A numeric value for the initial estimate of the intercompartmental clearance (Q). Default is 1.}

\item{input.add}{A numeric value for the additive error model. Default is 1.}
}
\value{
A list containing the following elements:
\item{npd.2cmpt_results}{A data frame with the estimated CL, Vc, Vp, Q, and
the time spent during estimation.}
\item{npd.2cmpt.APE}{The absolute prediction error (APE).}
\item{npd.2cmpt.MAPE}{The mean absolute percentage error (MAPE).}
\item{npd.list.2cmpt}{The full output list from the \code{Fit_2cmpt_oral} function.}
}
\description{
Perform parameter estimation with naive pooled data approach analysis for a two-compartment oral model, and evaluate the model fitting performance by the absolute prediction error (APE) and mean absolute prediction error (MAPE).
}
\details{
This function fits the two-compartment IV model to the given dataset using
the specified estimation method. It excludes rows where \code{EVID == 2}, as these
represent non-observation events. After fitting the model, the function returns
the estimated CL, Vc, Vp, and Q along with the APE (absolute prediction error)
and MAPE (mean absolute percentage error) to assess model fit.

The function also applies penalties to the results if certain conditions are
met, such as high relative standard error (\verb{\%RSE > 50}) or if the fitting method
does not converge successfully for the \code{focei} method.
}
\examples{
\dontrun{
# Example usage:
result <- run_npd_2cmpt_oral(dat = Oral_2CPT,
                           input.ka = 1,
                           input.cl = 4,
                           input.vc2cmpt = 35,
                           input.vp2cmpt = 35,
                           input.q2cmpt = 4)
}

}
