% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modelfit.R
\name{run_npd_3cmpt_iv}
\alias{run_npd_3cmpt_iv}
\title{Run and evaluate three-compartment IV Model}
\usage{
run_npd_3cmpt_iv(
  dat,
  est.method = "nls",
  input.cl = exp(1),
  input.vc3cmpt = exp(1),
  input.vp3cmpt = exp(1),
  input.vp23cmpt = exp(1),
  input.q3cmpt = exp(1),
  input.q23cmpt = exp(1),
  input.add = 1
)
}
\arguments{
\item{dat}{A data frame containing intravenous pharmacokinetic data. It should include
columns such as \code{ID}, \code{EVID}, \code{DV}, \code{dose}, and \code{AMT} and other necessary variables.}

\item{est.method}{The estimation method in nlmixr2 to use (e.g., "nls", "nlm", focei"...). The default value is "nls".}

\item{input.cl}{A numeric value for the initial estimate of log-transformed clearance (CL). Default is exp(1).}

\item{input.vc3cmpt}{A numeric value for the initial estimate of the volume of distribution in the central compartment (Vc). Default is exp(1).}

\item{input.vp3cmpt}{A numeric value for the initial estimate of the volume of distribution in the peripheral compartment (Vp). Default is exp(1).}

\item{input.vp23cmpt}{A numeric value for the initial estimate of the volume of distribution in the peripheral compartment (Vp2). Default is exp(1).}

\item{input.q3cmpt}{A numeric value for the initial estimate of the intercompartmental clearance (Q). Default is exp(1).}

\item{input.q23cmpt}{A numeric value for the initial estimate of the intercompartmental clearance (Q2). Default is exp(1).}

\item{input.add}{A numeric value for the additive error model. Default is 1.}
}
\value{
A list containing the following elements:
\item{npd.3cmpt_results}{A data frame with the estimated CL, Vc, Vp, Q, and
the time spent during estimation.}
\item{npd.3cmpt.APE}{The absolute prediction error (APE).}
\item{npd.3cmpt.MAE}{The mean absolute error (MAE).}
\item{npd.3cmpt.MAPE}{The mean absolute percentage error (MAPE).}
\item{npd.3cmpt.RMSE}{The root mean square error (RMSE).}
\item{npd.3cmpt.rRMSE}{The relative root mean square error (rRMSE).}
\item{npd.list.3cmpt}{The full output list from the \code{Fit_3cmpt_iv} function.}
}
\description{
Perform parameter estimation with naive pooled data approach analysis for a three-compartment intravenous model, and evaluate the model fitting performance by the absolute prediction error (APE) and mean absolute prediction error (MAPE).
}
\details{
This function fits the three-compartment IV model to the given dataset using
the specified estimation methods. It excludes rows where \code{EVID == 2}.
After fitting the model, the function returns the estimated CL and Vd along
with metrics to assess model fit.
}
\examples{
\dontrun{
# Example usage:
result <- run_npd_3cmpt_iv(dat = Bolus_2CPT,
                           input.cl = 4,
                           input.vc3cmpt = 10,
                           input.vp3cmpt = 10,
                           input.vp23cmpt = 10,
                           input.q3cmpt = 4,
                           input.q23cmpt = 4
                           )
}

}
