% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simpcal_iv.R
\name{run_simpcal_iv}
\alias{run_simpcal_iv}
\title{Run approximate pharmacokinetic calculation for intravenous pharmacokinetic data}
\usage{
run_simpcal_iv(dat, half_life = NA)
}
\arguments{
\item{dat}{A data frame containing the intravenous pharmacokinetic data.}

\item{half_life}{The half-life of the drug.}
}
\value{
A list containing the results of the simplified calculation, including clearance, volume of distribution, and the processed dataset.
}
\description{
Perform a simplified calculation for clearance and volume of distribution in intravenous cases.
}
\examples{

# Example 1 (iv case)
dat <- Bolus_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)[1]
run_simpcal_iv(fdat,half_life)

# Example 2 (infusion case).

dat <- Infusion_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)[1]
run_simpcal_iv(fdat,half_life)
}
