% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sens_params.R
\name{sim_sens_vmax_km}
\alias{sim_sens_vmax_km}
\title{Simulation-based sensitivity analysis for Vmax and Km}
\usage{
sim_sens_vmax_km(
  dat,
  sim_vmax_list = NA,
  sim_km_list = NA,
  estvd,
  estcl,
  estcmax,
  estka = NA,
  noniv_flag = 0
)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data with columns for EVID and DV.}

\item{sim_vmax_list}{A list of Vmax values to simulate (optional).}

\item{sim_km_list}{A list of Km values to simulate (optional).}

\item{estcl}{Estimated clearance (optional, required if \code{sim_vmax_list} and \code{sim_km_list} are not provided).}

\item{estcmax}{Estimated maximum concentration (optional, required if \code{sim_vmax_list} and \code{sim_km_list} are not provided).}

\item{estvc}{Estimated volume of distribution.}
}
\value{
A data frame containing the Vmax, Km, APE, MAPE, and time spent for each simulation.
}
\description{
This function performs sensitivity analysis by testing a series of Vmax and Km values.
}
\examples{
dat <- Bolus_1CPT
sim_sens_vmax_km(dat = Bolus_1CPT,sim_vmax_list = c(500,1000,1500),sim_km_list = c(200,300,400),estvc = 70)
}
