% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_point.R
\name{single_point_base}
\alias{single_point_base}
\title{Perform basic single-point pharmacokinetic calculations}
\usage{
single_point_base(dat, half_life = NA)
}
\arguments{
\item{dat}{A data frame containing pharmacokinetic data, including information on doses, sampling times, and observed concentrations.}

\item{half_life}{The half-life of the drug, used for determining elimination phase parameters and steady-state conditions.}
}
\value{
A list containing the following:
\itemize{
\item \code{singlepoint.results}: A data frame with the calculated clearance (CL) and volume of distribution (Vd), along with processing start and elapsed time.
\item \code{dat}: The processed dataset with additional columns for steady-state and other calculated flags.
\item \code{dat.ss.obs}: Subset of the data used for steady-state clearance calculation.
\item \code{dat.fd.obs}: Subset of the data used for single-dose volume of distribution calculation.
}
}
\description{
Performs pharmacokinetic calculations using a single-point method and calculates clearance (CL) and volume of distribution (Vd) based on the provided pharmacokinetic data.
}
\details{
The function includes two main calculations:
\enumerate{
\item \strong{Clearance (CL):}
\itemize{
\item Calculated using steady-state data from multiple dosing regimens.
\item Steady-state concentrations are identified and used to estimate clearance based on the dose, observed steady-state concentrations, and dosing intervals.
}
\item \strong{Volume of Distribution (Vd):}
\itemize{
\item Calculated for intravenous single-dose data.
\item The first sampling point (C0) after dose administration is used to estimate Vd with the formula \code{Dose / C0}.
}
}

The function supports both bolus and short-term infusion routes. It processes input data to identify steady-state and first-dose conditions and calculates trimmed mean values for CL and Vd to reduce the impact of outliers.
}
\examples{
# Example 1 (iv case)
dat <- Bolus_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)[1]
run_single_point(fdat,half_life)

# Example 2 (infusion case).

dat <- Infusion_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)$half_life_median
single_point_base(fdat,half_life)
run_single_point(fdat,half_life)

 # Example 3 (Oral case).

dat <- Oral_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)$half_life_median
run_single_point(fdat,half_life)


}
