% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_nca.normalised.R
\name{trap.rule}
\alias{trap.rule}
\title{Calculate Area Under the Curve (AUC) using the trapezoidal rule}
\usage{
trap.rule(x, y)
}
\arguments{
\item{x}{A numeric vector representing the time points.}

\item{y}{A numeric vector representing the corresponding concentration at each time point.}
}
\value{
A numeric value representing the estimated AUC using the trapezoidal rule.
}
\description{
This function computes the area under the curve (AUC) using the linear trapezoidal rule.
}
\examples{
x <- c(0.5, 1, 2, 4, 6, 8, 10)
y <- c(12, 8, 5, 3, 2, 1.5, 1)
trap.rule(x, y)

}
