% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_point_base.R
\name{calculate_vd}
\alias{calculate_vd}
\title{Estimate volume of distribution from concentration Data}
\usage{
calculate_vd(
  dat,
  half_life = NULL,
  dose_type = NULL,
  pooled_ctrl = pooled_control(),
  route = c("bolus", "oral", "infusion")
)
}
\arguments{
\item{dat}{A data frame containing pharmacokinetic data. Must include at least
the following columns: EVID, dose_number, tad, iiobs, TIME, DV, and optionally
durationobs and rateobs for infusion routes.}

\item{half_life}{Optional numeric value for the drug's half-life.
If not provided, it will be estimated using get_hf() from pooled
observations.}

\item{dose_type}{Type of analysis to perform. One of:
\itemize{
\item first_dose: First-dose analysis
\item repeated_doses: Analysis of doses beyond the first
\item combined_doses: Analysis combining first and repeated doses
}}

\item{pooled_ctrl}{Optional list of control parameters used by get_pooled_data()
for pooling observations. Defaults to output from pooled_control().}

\item{route}{Character string specifying the route of administration. Must be
one of bolus, oral, or infusion. Currently, oral is not implemented.}
}
\value{
A named list with two components:
\describe{
\item{vd_df}{A \code{data.frame} of selected observations with an added \code{vd} column representing
the calculated volume of distribution for each subject.}
\item{trimmed_mean_vd}{A numeric value giving the trimmed geometric mean of individual Vd estimates,
using a 5\% trimming level to reduce the impact of outliers.}
}
}
\description{
Calculates the apparent volume of distribution (\eqn{V_d}) for intravenous
bolus or infusion routes using an adaptive single-point method. If the
half-life is not provided, it is estimated from pooled data.
}
\details{
The function uses an early concentration point (based on time since dosing and half-life)
to estimate Vd. For bolus:
\deqn{Vd = \frac{\text{Dose}}{C_0}}
For infusion:
\deqn{Vd = \frac{\text{Rate} \times \min(\text{TIME}, \text{durationobs})}{C}}
where \code{C} is the observed concentration (\code{DV}).

When \code{half_life} is not provided, it is estimated from pooled data using
\code{get_pooled_data()} and \code{get_hf()}.
}
\examples{
\dontrun{
# Bolus example
dat <- Bolus_1CPT
out <- processData(dat)
fdat<- out$dat
froute <-out$Datainfo$Value[out$Datainfo$Infometrics == "Dose Route"]
half_life <- get_hf(dat = fdat)$half_life_median
calculate_vd(dat = fdat, half_life = half_life,route=froute)

# Infusion example
dat <- Infusion_1CPT
out <- processData(dat)
fdat<- out$dat
froute <-out$Datainfo$Value[out$Datainfo$Infometrics == "Dose Route"]
half_life <- get_hf(dat = fdat)$half_life_median
calculate_vd(dat = fdat, half_life = half_life,route=froute)

}
}
