% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{eval_perf_1cmpt}
\alias{eval_perf_1cmpt}
\title{Evaluate model performance for one-compartment PK model}
\usage{
eval_perf_1cmpt(
  dat,
  est.method = "rxSolve",
  ka = NULL,
  cl = NULL,
  vd = NULL,
  route = c("bolus", "infusion", "oral")
)
}
\arguments{
\item{dat}{A data.frame containing PK data with columns such as \code{EVID} and \code{DV}.}

\item{est.method}{Estimation method passed to the fitting function.
Defaults to using \code{rxSolve} for model simulation and parameter estimation.
Alternative methods can be specified as needed (e.g., "nls", "nlm","nlminb").}

\item{ka}{Absorption rate constant (only used for oral route).}

\item{cl}{Clearance value.}

\item{vd}{Volume of distribution.}

\item{route}{A character string indicating the route of administration.
Must be one of \code{"oral"}, \code{"infusion"}, or \code{"bolus"}. Defaults to \code{"bolus"}.}
}
\value{
A numeric vector of length 5 containing the following performance metrics (rounded to 3 decimal places):
\describe{
\item{APE}{Absolute prediction error}
\item{MAE}{Mean absolute error}
\item{MAPE}{Mean absolute percentage error}
\item{RMSE}{Root mean square error}
\item{rRMSE}{Relative RMSE}
}
If simulation fails or input parameters are missing, returns a vector of NAs.
}
\description{
Computes common predictive performance metrics (e.g., APE, MAE, MAPEM, RMSE, rRMSE)
for a pharmacokinetic (PK) model based on simulated and observed concentration data.
}
\details{
Internally selects the appropriate one-compartment model fitting function depending on the route of administration:
\itemize{
\item Uses \code{Fit_1cmpt_oral()} for oral data
\item Uses \code{Fit_1cmpt_iv()} for intravenous data
}
}
\examples{
\dontrun{
eval_perf_1cmpt(
  dat = Oral_1CPT,
  est.method = "rxSolve",
  ka = 1,
  cl = 4,
  vd = 70,
  route = "oral"
)
}
}
\seealso{
\code{\link{Fit_1cmpt_oral}}, \code{\link{Fit_1cmpt_iv}}, \code{\link{metrics.}}
}
