% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npd_modelcodes.R
\name{Fit_1cmpt_mm_oral}
\alias{Fit_1cmpt_mm_oral}
\title{Fit oral pharmacokinetic data to a one-compartment model with Michaelis-Menten Kinetics}
\usage{
Fit_1cmpt_mm_oral(
  data,
  est.method,
  input.ka,
  input.vmax,
  input.km,
  input.vd,
  input.add
)
}
\arguments{
\item{data}{Intravenous pharmacokinetic data in the nlmixr2 format.}

\item{est.method}{The estimation method in nlmixr2 to use (e.g., "nls", "nlm", focei"...). The default value is "nls".}

\item{input.ka}{Initial estimate for absorption rate constant.}

\item{input.vmax}{Initial estimate for Vmax.}

\item{input.km}{Initial estimate for Km.}

\item{input.vd}{Initial estimate for volume of distribution.}

\item{input.add}{Initial estimate for additive error.}
}
\value{
An object of class 'nlmixr2' containing the results of the estimation.
}
\description{
Perform parameter estimation using the naive pooled data approach with a one-compartment model with nonlinear elimination kinetics for the provided intravenous pharmacokinetic data.
}
\examples{
dat <- Oral_1CPTMM
Fit_1cmpt_mm_oral(dat=dat,est.method="nls",input.ka=1,input.vmax=1000,input.km=100,input.vd=70,input.add=1)
}
