% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npd_modelcodes.R
\name{Fit_1cmpt_oral}
\alias{Fit_1cmpt_oral}
\title{Fit oral pharmacokinetic data to a one-compartment model}
\usage{
Fit_1cmpt_oral(data, est.method, input.ka, input.cl, input.vd, input.add)
}
\arguments{
\item{data}{Intravenous pharmacokinetic data in the nlmixr2 format.}

\item{est.method}{The estimation method in nlmixr2 to use (e.g., "nls", "nlm", "focei"). The default value is "nls".}

\item{input.ka}{Initial estimate for absorption rate constant.}

\item{input.cl}{Initial estimate for clearance.}

\item{input.vd}{Initial estimate for volume of distribution.}

\item{input.add}{Initial estimate for additive error.}
}
\value{
An object of class 'nlmixr2' containing the results of the estimation.
}
\description{
Perform parameter estimation using the naive pooled data approach with a one-compartment model for the provided intravenous pharmacokinetic data.
}
\examples{
dat <- Oral_1CPT
Fit_1cmpt_iv(dat=dat,est.method="nls",input.ka=1,input.cl=4,input.vd=70,input.add=1)
}
