% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npd_modelcodes.R
\name{Fit_2cmpt_iv}
\alias{Fit_2cmpt_iv}
\title{Fit intravenous pharmacokinetic data to a two-compartment model}
\usage{
Fit_2cmpt_iv(
  data,
  est.method,
  input.cl,
  input.vc2cmpt,
  input.vp2cmpt,
  input.q2cmpt,
  input.add
)
}
\arguments{
\item{data}{Intravenous pharmacokinetic data in the nlmixr2 format.}

\item{est.method}{The estimation method in nlmixr2 to use (e.g., "nls", "nlm", focei"...). The default value is "nls".}

\item{input.cl}{Initial estimate for clearance.}

\item{input.vc2cmpt}{Initial estimate for the central volume of distribution in two-compartment model.}

\item{input.vp2cmpt}{Initial estimate for the peripheral volume of distribution in a two-compartment model.}

\item{input.q2cmpt}{Initial estimate for inter-compartmental clearance in a two-compartment model.}

\item{input.add}{Initial estimate for additive error.}
}
\value{
An object of class 'nlmixr2' containing the results of the estimation.
}
\description{
Perform parameter estimation using the naive pooled data approach with a one-compartment model for the provided intravenous pharmacokinetic data.
}
\examples{
dat <- Bolus_2CPT
Fit_2cmpt_iv(dat=dat,est.method="nls",input.cl=4,input.vc2cmpt=70,input.vp2cmpt=35,input.q2cmpt=10,input.add=1)
}
