% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getppkinit.R
\name{getppkinit}
\alias{getppkinit}
\title{Get initial estimates for a population pharmacokinetic modelling}
\usage{
getppkinit(dat, runnpd, getinit.settings)
}
\arguments{
\item{dat}{A data frame containing the intravenous pharmacokinetic data.}

\item{runnpd}{An integer indicating whether to run the naive pooled data approach (default is 0).}

\item{getinit.settings}{A list or data frame containing calculation settings (optional). It should include 'half_life', 'nlastpoints', 'nbins', and 'est.method'.}
}
\value{
A list containing data information, initial parameter estimates, messages, and run history.
}
\description{
Calculates initial pharmacokinetic parameters using multiple methods including simplified calculation, non-compartmental analysis, graphical methods, and parameter estimation with naive pooled data approach if specified.
}
\examples{
getppkinit(dat = Bolus_1CPT,runnpd = 0)
getppkinit(dat = Bolus_1CPT,runnpd = 1)
}
