% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_graphcal.R
\name{graphcal_iv}
\alias{graphcal_iv}
\title{Graphical calculation of clearance and volume of distribution}
\usage{
graphcal_iv(dat, nlastpoints = 3)
}
\arguments{
\item{dat}{A data frame with intravenous pharmacokinetic data, at least containing TIME and DV these two columns}

\item{nlastpoints}{Numeric value specifying the number of last data points used for linear regression to obtain the slope in the terminal phase. (default is \code{3} in the graphical analysis).}
}
\value{
A named vector containing the calculated clearance (cl), volume of distribution (vd), slope of terminal phase (slope), and concentration at the time 0 (C0).
}
\description{
Graphic calculation for clearance and volume of distribution from the provided data.
}
\examples{
dat <- data.frame(TIME = c(0.5, 1, 2, 4, 6, 8, 10), DV = c(12, 8, 5, 3, 2, 1.5, 1 ))
graphcal_iv(dat, nlastpoints = 3)
}
