% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_graphcal.R
\name{graphcal_oral}
\alias{graphcal_oral}
\title{Graphical calculation of clearance and volume of distribution for oral case}
\usage{
graphcal_oral(dat, nlastpoints)
}
\arguments{
\item{dat}{A data frame with oral pharmacokinetic data, at least containing TIME and DV these two columns}

\item{nlastpoints}{Number of last points to use for the linear regression of terminal slope}
}
\value{
A named vector containing the calculated clearance (cl), volume of distribution (vd), slope of terminal phase (slope), and concentration at the time 0 (C0).
}
\description{
Graphic calculation for clearance and volume of distribution from the provided data.
}
\examples{
dat <- data.frame(TIME = c(0.5, 1, 2, 4, 6, 8, 10), DV = c(1, 2, 5, 3, 2, 1.5, 1 ))
graphcal_oral(dat, nlastpoints = 3)
}
