% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/half_life_estimated.R
\name{half_life_estimated}
\alias{half_life_estimated}
\title{Estimate half-life from given pharmacokinetic Data}
\usage{
half_life_estimated(dat, sdflag, fdobsflag, nlastpoints)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data. It should include columns
for \code{DV} (dependent variable), \code{dose}, and \code{dose_number}.}

\item{sdflag}{A numeric flag indicating if the data includes repeated dose samples.
\code{0} means repeated dose data is included, \code{1} otherwise.}

\item{fdobsflag}{A numeric flag indicating if the data includes only the first dose samples.
\code{1} means only first dose data is included, \code{0} otherwise.}

\item{nlastpoints}{An integer specifying the number of last data points to be used for
linear regression in the estimation.}
}
\value{
A numeric value representing the geometric mean of the estimated half-life
values from the different data subsets (first dose, repeated doses, and all data).
}
\description{
Estimates the half-life based on pharmacokinetic data by linear regression on terminal phase,
considering different scenarios: data after the first dose, data after repeated doses,
and data including both first and repeated doses.
}
\details{
The function splits the input data into subsets based on the dose number,
performs time binning, and estimates the half-life using the slope of the log-linear
regression on the last points of the normalized concentration-time data.

The half-life is estimated separately for:
\itemize{
\item Only first dose data
\item Only repeated dose data
\item All data combined
}
The final result is the geometric mean of the positive half-life estimates from these
subsets.
}
\examples{
# Example usage:
dat <- Bolus_1CPT
dat <- nmpkconvert(dat)
dat <- calculate_tad(dat)
half_life_estimated(dat, sdflag = 0, fdobsflag = 1, nlastpoints = 4)

}
