% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initsControl.R
\name{initsControl}
\alias{initsControl}
\title{Create full control list for initial PPK parameter estimation}
\usage{
initsControl(
  ss.control = ss_control(),
  pooled.control = pooled_control(),
  nca.control = nca_control(),
  fallback.control = fallback_control(),
  selection.criterion = "All"
)
}
\arguments{
\item{pooled_control}{Output of \code{pooled_control()}.}

\item{nca_control}{Output of \code{nca_control()}.}

\item{selection_control}{Output of \code{selection_control()}.}

\item{fallback_control}{Output of \code{fallback_control()}.}
}
\value{
A named list combining all control modules.
}
\description{
Aggregates all modular control functions into one structured list.
}
\examples{
initsControl(
  pooled_control = pooled_control(nbins = 8),
  fallback_control = fallback_control(sigma_method_additive = "fixed_fraction")
)
}
