% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ss.R
\name{is_ss}
\alias{is_ss}
\title{Identify steady-state observations}
\usage{
is_ss(df, half_life, dose_interval)
}
\arguments{
\item{df}{A data frame containing the pharmacokinetic data. The data frame must include columns for ID, TIME, EVID, AMT, and tad (time after dose).}

\item{half_life}{The half-life of the drug.}

\item{dose_interval}{The dosing interval.}
}
\value{
A data frame with an additional column \code{SteadyState} indicating whether each observation is at steady state.
}
\description{
Checks if a sufficient number of doses have been administered before each observation point to determine if the observation point has reached a steady state, and calculates the time required to reach steady state based on the provided half-life and dose interval, and marks observations as steady state if they meet the criteria.
}
\examples{
df <- data.frame(ID = rep(1, 10), EVID = c(1, 0, 0, 1, 0, 0, 1, 0, 0, 1), TIME = 1:10, AMT = c(100, 0, 0, 100, 0, 0, 100, 0, 0, 100), tad = runif(10))
is_ss(df, half_life = 4, dose_interval = 8)
}
