% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_ss.R
\name{is_ss}
\alias{is_ss}
\title{Determine steady state for pharmacokinetic observations}
\usage{
is_ss(df, ssctrl = ss_control(), half_life = NA)
}
\arguments{
\item{df}{A data frame containing pharmacokinetic data. It should include columns:
\itemize{
\item \code{ID}: Unique identifier for each subject.
\item \code{EVID}: Event ID, where 1 indicates a dose and 0 indicates an observation.
\item \code{SSflag}: Steady-state flag, set to 1 if steady state is manually indicated.
\item \code{TIME}: Observation or dosing time.
\item \code{AMT}: Dose amount.
\item \code{tad}: Time after the last dose.
}}

\item{ssctrl}{A list created by \code{\link{ss_control}()}, specifying control settings
for steady-state evaluation. See \code{\link{ss_control}} for details on available options.}

\item{half_life}{Numeric. The drug's half-life. Required if \code{ss_method} in \code{ssctrl} is
\code{"combined"} or \code{"half_life_based"}.}
}
\value{
A data frame similar to \code{df} with additional columns:
\itemize{
\item \code{SteadyState}: Logical, indicating whether each observation has reached steady state.
\item \code{recent_ii}: Numeric, the most recent dosing interval for observations marked as steady state.
\item \code{SS.method}: Character, description of the method used to identify steady state.
}
}
\description{
Evaluates whether observations in a pharmacokinetic dataset have reached steady state,
based on user-defined criteria. It supports three methods: using a fixed number of doses,
using a fixed number of half-lives, or combining both criteria.
}
\examples{
# Example usage
dat <- pheno_sd
dat <- processData(dat)$dat
out<-is_ss(df = dat)
out[out$SteadyState == TRUE & !is.na(out$SteadyState),
           c("ID", "TIME", "DV", "EVID", "SteadyState")]
}
