% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_ka.R
\name{ka_calculation_md}
\alias{ka_calculation_md}
\title{Calculate absorption rate constant (ka) in a multiple-dose one-compartment model}
\usage{
ka_calculation_md(cl, ke, t, Ct, Fbio = 1, Dose, tau)
}
\arguments{
\item{cl}{Numeric. Clearance of the drug (in L/hr).}

\item{ke}{Numeric. Elimination rate constant (in 1/hr).}

\item{t}{Numeric. Time after the last dose (in hours) at which the concentration is measured.}

\item{Ct}{Numeric. Observed concentration of the drug at time \code{t} (in mg/L).}

\item{Fbio}{Numeric. Bioavailability fraction (default = 1, meaning 100\% bioavailability).}

\item{Dose}{Numeric. Administered dose of the drug (in mg).}

\item{tau}{Numeric. Dosing interval (in hours) between successive doses.}
}
\value{
A list containing the following components:
\item{ka}{The calculated absorption rate constant.}
\item{full_solution}{The full solution object returned by the \code{uniroot()} function, which includes additional details about the root-finding process.}
}
\description{
Estimates the absorption rate constant (\code{ka}) for a drug administered orally
in a multiple-dose regimen using a one-compartment model with first-order absorption
and elimination kinetics. The calculation is based on known values of clearance (\code{cl}),
elimination rate constant (\code{ke}), time since last dose (\code{t}), observed concentration
(\code{Ct}), bioavailability (\code{Fbio}), dose (\code{Dose}), dosing interval (\code{tau}),
and the number of doses administered (\code{n}).
}
\details{
This function uses a multiple-dose one-compartment model with first-order absorption and
elimination to estimate the absorption rate constant (\code{ka}). The function solves the
following equation for \code{ka} numerically using \code{uniroot()}:

\deqn{Ct = \frac{Fbio \cdot Dose \cdot ka}{Vd \cdot (ka - ke)} \left( \frac{e^{-ke \cdot t}}{1 - e^{-ke \cdot \tau}} - \frac{e^{-ka \cdot t}}{1 - e^{-ka \cdot \tau}} \right)}

The \code{uniroot()} function is used to find the value of \code{ka} that makes the difference
between the predicted and observed concentrations equal to zero. The reasonable range for
\code{ka} is set as \link{0.01, 1000} for root-finding.
}
\examples{
# Example usage:
ka_result <- ka_calculation_md(cl = 4, ke = 0.057, t = 2, Ct = 852, Dose = 60000, tau = 24)
ka_result <- ka_calculation_md(cl = 4, ke = 0.057, t = 2, Ct = 189.6, Dose = 10000, tau = 24)
print(ka_result$ka)

}
