% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_ka.R
\name{ka_wanger_nelson}
\alias{ka_wanger_nelson}
\title{Calculate the Absorption Rate Constant (ka) using the Wanger-Nelson Method}
\usage{
ka_wanger_nelson(dat, nca.out = NULL)
}
\arguments{
\item{dat}{A data frame with at least two columns: \code{DV} (drug concentration)
and \code{TIME} (time points of measurement).}

\item{nca.out}{(Optional) An object containing results from a prior NCA
computation (e.g., output from \code{getnca()}). If not provided, the function
internally calls \code{getnca()} to estimate AUC and the elimination rate
constant (\code{ke}). Providing \code{nca.out} can save computation time when
NCA results are already available.}
}
\value{
A numeric value representing the estimated absorption rate constant (ka).
}
\description{
Computes the absorption rate constant (ka) using the Wanger-Nelson method.
The function makes use of non-compartmental analysis (NCA) function part to calculate the area under the curve (AUC) and the elimination rate constant (ke).
Linear regression on the logarithm of the remaining fraction of drug is used to estimate the absorption rate constant (ka).
}
\details{
The function calculates the absorption rate constant (ka) as follows:
\itemize{
\item The function uses the \code{nca.iv.normalised} method to calculate the area under the curve (AUC) and elimination rate constant (ke).
\item AUC is calculated using trapezoidal integration for the observed data points.
\item The fraction of drug absorbed is calculated based on ke and the cumulative AUC.
\item The logarithm of the remaining fraction of the drug is then modeled using linear regression to determine the absorption rate constant (ka).
}

This method assumes that the drug absorption follows first-order kinetics.
}
\examples{

# Calculate ka for the first subject from Oral_1CPT
dat<-Oral_1CPT[Oral_1CPT$ID==1 &Oral_1CPT$SD==1&Oral_1CPT$EVID==0,]
dat<-data.frame(TIME=dat$TIME,DV=dat$DV)
ka_wanger_nelson(dat=dat)$ka

}
