% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tad.R
\name{mark_dose_number}
\alias{mark_dose_number}
\title{Mark dose number}
\usage{
mark_dose_number(dat)
}
\arguments{
\item{dat}{A data frame containing raw time–concentration data in the
standard nlmixr2 format.}
}
\value{
A modified data frame with an added column named dose_number, indicating
the sequential dose count within each subject and reset group.
}
\description{
Assigns sequential dose numbers based on dosing events (EVID) within each subject.
}
\examples{
mark_dose_number(Bolus_1CPT)
mark_dose_number(Infusion_1CPT)
mark_dose_number(Oral_1CPT)

}
\author{
Zhonghui Huang
}
