% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tad.R
\name{mark_dose_number}
\alias{mark_dose_number}
\title{Mark dose number}
\usage{
mark_dose_number(dat)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data. The data frame must include columns for ID, TIME, EVID, and AMT.}
}
\value{
A data frame with an additional column \code{dose_number} indicating the sequential number of each dose event for each subject.
}
\description{
Mark the dose number in the dataset by assigning a sequential dose number identified by EVID to each dose event.
}
\examples{
dat <-Bolus_1CPT
mark_dose_number(dat)
}
