% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pooled_nca.R
\name{nca_control}
\alias{nca_control}
\title{Control settings for non-compartmental analysis (NCA)}
\usage{
nca_control(
  trapezoidal.rule = c("linear_up_log_down", "linear"),
  duration = NULL,
  nlastpoints = 3
)
}
\arguments{
\item{nlastpoints}{Integer. Number of terminal points for half-life regression.}

\item{trapezoidal_rule}{Integer. 1 = linear, 2 = linear-up log-down (default = 1).}

\item{half_life}{Numeric or NA. Optional fixed half-life override.}
}
\value{
A list with NCA control parameters.
}
\description{
Control settings for non-compartmental analysis (NCA)
}
\examples{
nca_control()
}
