% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pooled_nca.R
\name{nca_control}
\alias{nca_control}
\title{Control settings for non-compartmental analysis (NCA)}
\usage{
nca_control(
  trapezoidal.rule = c("linear_up_log_down", "linear"),
  duration = NULL,
  nlastpoints = 3,
  slope.method = "bestfitforce"
)
}
\arguments{
\item{duration}{Numeric. Optional. Duration of the observation window (same units as time).
Used to restrict the integration or define the evaluation range.}

\item{nlastpoints}{Integer. Number of terminal points for half-life regression (default = 3).}

\item{trapezoidal_rule}{Character. Method for trapezoidal AUC integration:
\itemize{
\item \code{"linear"} - Linear trapezoidal rule (default)
\item \code{"linear_up_log_down"} - Linear-up / log-down rule
}}

\item{slope_method}{Character. Method for estimating the terminal slope (\eqn{\lambda_z}):
\itemize{
\item \code{"bestfitforce"} - Force estimation using decreasing number of terminal points if best-fit fails (default)
\item \code{"bestfit"} - Use automated best-fit selection based on adjusted R-squared
}}
}
\value{
A list with NCA control parameters.
}
\description{
Defines control parameters used in non-compartmental pharmacokinetic analysis (NCA),
including numerical integration method, duration, number of terminal points, and
slope estimation strategy.
}
\examples{
nca_control()
}
