% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nmpkconvert.R
\name{nmpkconvert}
\alias{nmpkconvert}
\title{Convert pharmacokinetic dataset to depreciated format with additional dosing}
\usage{
nmpkconvert(dat)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data with columns for EVID, ADDL, TIME, II, AMT, and other relevant variables.}
}
\value{
A data frame with expanded dosing records and sorted by ID and TIME.
}
\description{
Process pharmacokinetic (PK) data to handle additional dosing (ADDL) by expanding the dose records accordingly.
}
\examples{
dat <- data.frame(ID = rep(1, 6), EVID = c(1, 0, 0, 1, 0, 0), ADDL = c(2, 0, 0, 0, 0, 0), TIME = c(0, 1, 2, 3, 4, 5), II = c(24, 0, 0, 0, 0, 0), AMT = c(100, 0, 0, 0, 0, 0))
nmpkconvert(dat)
}
