% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_graphcal.R
\name{run_graphcal}
\alias{run_graphcal}
\title{Run graphical calculation}
\usage{
run_graphcal(dat, fdobsflag, noniv_flag, nbins, nlastpoints)
}
\arguments{
\item{dat}{A data frame containing the intravenous pharmacokinetic data.}

\item{fdobsflag}{A flag indicating whether the dataset includes first-dose observations. Set to 1 if the data includes first-dose observations, otherwise set to 0.}

\item{nbins}{The number of bins for time binning.}

\item{nlastpoints}{The number of last points to be used for slope linear regression.}
}
\value{
A data frame with calculated clearance, volume of distribution, slope, and time spent for the calculation.
}
\description{
Performs a graphical calculation of clearance and volume of distribution based on provided pharmacokinetic data with concentration normalised by dose
}
\examples{
# example 1 (iv case)

dat <- Bolus_1CPT
dat <- nmpkconvert(dat)
dat <- calculate_tad(dat)
run_graphcal(dat, fdobsflag = 1, noniv_flag=0, nbins = 8, nlastpoints = 4)

# example2 (oral case)
dat <- Oral_1CPT
dat <- nmpkconvert(dat)
dat <- calculate_tad(dat)
run_graphcal(dat, fdobsflag = 1,  noniv_flag=1, nbins = 8, nlastpoints = 4)
}
