% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_graphcal.R
\name{run_graphcal}
\alias{run_graphcal}
\title{Graphical calculation of pharmacokinetic parameters}
\usage{
run_graphcal(dat, route, nbins = 8, nlastpoints = 3, fdobsflag = 0)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data.}

\item{route}{A character string specifying the route of administration. It can be "bolus",
"infusion", or "oral". The function will adapt its calculations based on the specified route.}

\item{nbins}{An integer specifying the number of bins used to divide the time points for naive pooling of the data. The default is 8.}

\item{nlastpoints}{Numeric value specifying the number of last data points used for linear regression to obtain the slope in the terminal phase. (default is \code{3} in the graphical analysis).}
}
\value{
A data frame with calculated clearance, volume of distribution, slope, and time spent for the calculation.
}
\description{
Performs a graphical calculation of clearance and volume of distribution based on provided pharmacokinetic data with concentration normalised by dose
}
\examples{

# Example 1 (iv case)
dat <- Bolus_1CPT
dat <- nmpkconvert(dat)
dat <- calculate_tad(dat)
run_graphcal(dat, route="bolus")

# Example 2 (oral case)
dat <- Oral_1CPT
dat <- nmpkconvert(dat)
dat <- calculate_tad(dat)
run_graphcal(dat, route="oral")

# Example 3 (infusion case).
Approximate calculation. only use when the infusion duration is very short

dat <- Infusion_1CPT
dat <- nmpkconvert(dat)
dat <- calculate_tad(dat)
run_graphcal(dat, route="infusion")

}
