% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_graphcal.R
\name{run_graphcal}
\alias{run_graphcal}
\title{Run graphical analysis of PK Parameters}
\usage{
run_graphcal(dat, route, data_type = "first_dose", pooled = NULL, ...)
}
\arguments{
\item{dat}{A data frame containing pharmacokinetic data.}

\item{route}{Administration route, must be one of \code{"bolus"}, \code{"infusion"}, or \code{"oral"}.}

\item{data_type}{Type of dataset to use for binning (default: \code{"first_dose"}).}

\item{pooled}{Optional externally provided pooled data. If \code{NULL}, the data will be pooled internally.}

\item{...}{Additional arguments passed to \code{bin.time} or to the graphical calculation functions.}
}
\value{
A list containing graphical estimation results, including clearance, volume of distribution, terminal slope,
extrapolated concentration, and for oral data, absorption rate constant.
}
\description{
Performs graphical estimation of pharmacokinetic parameters based on the input data and route of administration,
returning detailed results in a list format.
}
\examples{
# Example 1 (iv case)
dat <- Bolus_1CPT
dat <- processData(dat)$dat
run_graphcal(dat, route="bolus")

# Example 2 (oral case)
dat <- Oral_1CPT
dat <- processData(dat)$dat
run_graphcal(dat, route="oral")

# Example 3 (infusion case).
Approximate calculation. only use when the infusion duration is very short

dat <- Infusion_1CPT
dat <- processData(dat)$dat
run_graphcal(dat, route="infusion")

}
\seealso{
\code{\link{graphcal_iv}}, \code{\link{graphcal_oral}}, \code{\link{get_pooled_data}}
}
