% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modelfit.R
\name{run_npd_1cmpt_mm_iv}
\alias{run_npd_1cmpt_mm_iv}
\title{Run and evaluate one-compartment IV model with Michaelis-Menten Kinetics
Perform parameter estimation on naive pooled data for a one-compartment Michaelis-Menten IV model.}
\usage{
run_npd_1cmpt_mm_iv(
  dat,
  est.method = "nls",
  npdmm_inputvmax = exp(1),
  npdmm_inputkm = exp(1),
  npdmm_inputcl = exp(1),
  npdmm_inputvd = exp(1),
  input.add = 1,
  km_threshold = F
)
}
\arguments{
\item{dat}{A data frame containing pharmacokinetic data. Must include columns
such as \code{ID}, \code{EVID}, \code{DV}, \code{dose}, and \code{AMT}.}

\item{est.method}{The estimation method in nlmixr2 to use (e.g., "nls", "nlm", focei"...). The default value is "nls".}

\item{npdmm_inputvmax}{A numeric value for the initial estimate of Vmax. Default is exp(1).}

\item{npdmm_inputkm}{A numeric value for the initial estimate of Km. Default is exp(1).}

\item{npdmm_inputcl}{A numeric value for the clearance. Default is exp(1).}

\item{npdmm_inputvd}{A numeric value for the initial estimate of volume of distribution (Vd). Default is exp(1).}

\item{km_threshold}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE},
initial estimates for \eqn{V_{max}} and \eqn{K_m} will be set based on the
observed maximum concentration in the dataset and referenced clearance.}
}
\value{
A list with the following elements:
\describe{
\item{npd.1cmpt.mm_results}{A data frame with the estimated values of \eqn{V_{max}},
\eqn{K_m}, and \eqn{V_d}, as well as the time taken for the estimation.}
\item{npd.1cmpt.mm..APE}{The absolute prediction error (APE).}
\item{npd.1cmpt.mm..MAE}{The mean absolute error (MAE).}
\item{npd.1cmpt.mm..MAPE}{The mean absolute percentage error (MAPE).}
\item{npd.1cmpt.mm..RMSE}{The root mean square error (RMSE).}
\item{npd.1cmpt.mm.rRMSE}{The relative root mean square error (rRMSE).}
\item{npd.1cmpt.mm.list}{The full output list from the \code{Fit_1cmpt_mm_iv} function.}
}
}
\description{
Estimates the pharmacokinetic parameters (Vmax, Km, Vd) for a
1-compartment Michaelis-Menten IV (Intravenous) model by fitting.
Optionally, an initial estimate of Km is set based on the
maximum concentration in the dataset (if \code{km_threshold = TRUE}).
}
\details{
If \code{km_threshold = TRUE}, this function first calculates initial estimates for \eqn{V_{max}} and \eqn{K_m}
based on the observed maximum concentration in the dataset and referenced clearance. This approach ensures
that the value of \eqn{K_m} is adjusted to lie between the linear and nonlinear regimes, providing a more robust
starting estimate when there is uncertainty about whether the model follows linear or nonlinear kinetics.
This adjustment ensures that the fitting process does not result in \eqn{K_m} values too far from reality,
regardless of whether the dynamics are linear or nonlinear.
}
\examples{
\dontrun{
# Example usage:
result <- run_npd_1cmpt_mm_iv(dat = Bolus_1CPT,
                              est.method="foce",
                              npdmm_inputcl = 4,
                              npdmm_inputvd = 70,
                              km_threshold = TRUE)
}

}
