% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modelfit.R
\name{run_npd_1cmpt_mm_oral}
\alias{run_npd_1cmpt_mm_oral}
\title{Run and evaluate one-compartment oral model with Michaelis-Menten Kinetics
Perform parameter estimation on naive pooled data for a one-compartment Michaelis-Menten oral model.}
\usage{
run_npd_1cmpt_mm_oral(
  dat,
  est.method = "nls",
  input.ka = exp(1),
  input.vmax = exp(1),
  input.km = exp(1),
  input.cl = exp(1),
  input.vd = exp(1),
  input.add = 1,
  km_threshold = F
)
}
\arguments{
\item{dat}{A data frame containing pharmacokinetic data. Must include columns
such as \code{ID}, \code{EVID}, \code{DV}, \code{dose}, and \code{AMT}.}

\item{est.method}{The estimation method in nlmixr2 to use (e.g., "nls", "nlm", focei"...). The default value is "nls".}

\item{input.ka}{A numeric value for the initial estimate of the log-transformed absorption rate constant. Default is 1.}

\item{input.add}{A numeric value for the additive error model. Default is Default is 1.}

\item{km_threshold}{A logical value (\code{TRUE} or \code{FALSE}). If \code{TRUE},}

\item{npdmm_inputvmax}{A numeric value for the initial estimate of the Vmax. Default is 1.}

\item{npdmm_inputkm}{A numeric value for the initial estimate of the Km. Default is 1.}

\item{npdmm_inputcl}{A numeric value for clearance for Vmax and Km calculation when km_threshold is set to True.}

\item{npdmm_inputvd}{A numeric value for the initial estimate of the volume of distribution (Vd). Default is 1.}
}
\value{
A list with the following elements:
\describe{
\item{npd.1cmpt.mm_results}{A data frame with the estimated values of  \eqn{K_a},\eqn{V_{max}},
\eqn{K_m}, and \eqn{V_d}, as well as the time taken for the estimation.}
\item{npd.1cmpt.mm..APE}{The absolute prediction error (APE).}
\item{npd.1cmpt.mm..MAE}{The mean absolute error (MAE).}
\item{npd.1cmpt.mm..MAPE}{The mean absolute percentage error (MAPE).}
\item{npd.1cmpt.mm..RMSE}{The root mean square error (RMSE).}
\item{npd.1cmpt.mm.rRMSE}{The relative root mean square error (rRMSE).}
\item{npd.1cmpt.mm.list}{The full output list from the \code{Fit_1cmpt_mm_oral} function.}
}
}
\description{
Estimates the pharmacokinetic parameters (Vmax, Km, Vd) for a
1-compartment Michaelis-Menten Oral (Intravenous) model by fitting.
Optionally, an initial estimate of Km is based on the
maximum concentration in the dataset (if \code{km_threshold = TRUE}).
}
\details{
If \code{km_threshold = TRUE}, this function first calculates initial estimates for \eqn{V_{max}} and \eqn{K_m}
based on the observed maximum concentration in the dataset and referenced clearance. This approach ensures
that the value of \eqn{K_m} is adjusted to lie between the linear and nonlinear regimes, providing a more robust
starting estimate when there is uncertainty about whether the model follows linear or nonlinear kinetics.
This adjustment ensures that the fitting process does not result in \eqn{K_m} values too far from reality,
regardless of whether the dynamics are linear or nonlinear.
}
\examples{
\dontrun{
# Example 1 (Linear kinetics case):
result <- run_npd_1cmpt_mm_oral(dat = Oral_1CPT,
                              est.method="foce",
                              input.cl = 4,
                              input.vd = 70,
                              km_threshold = TRUE)
result

# Example 2 ( nonlinear-kinetics case):
result2 <- run_npd_1cmpt_mm_oral(dat = Oral_1CPTMM,
                              input.cl = 4,
                              input.vd = 70,
                              km_threshold = TRUE)
result2
}

}
