% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_modelfit.R
\name{run_npd_3cmpt_iv}
\alias{run_npd_3cmpt_iv}
\title{Run and evaluate a three-compartment IV model}
\usage{
run_npd_3cmpt_iv(
  dat,
  est.method = "nls",
  input.cl = exp(1),
  input.vc3cmpt = exp(1),
  input.vp3cmpt = exp(1),
  input.vp23cmpt = exp(1),
  input.q3cmpt = exp(1),
  input.q23cmpt = exp(1),
  input.add = 1
)
}
\arguments{
\item{dat}{A data frame containing raw intravenous concentration–time data in
standard nlmixr2 format.}

\item{est.method}{Estimation method used in nlmixr2. Defaults to "nls".}

\item{input.cl}{Initial estimate for clearance (CL). Defaults to exp(1),
corresponding to a log-scale value of 1.}

\item{input.vc3cmpt}{Initial estimate for central volume of distribution (Vc).
Defaults to exp(1), corresponding to a log-scale value of 1.}

\item{input.vp3cmpt}{Initial estimate for first peripheral volume (Vp1).
Defaults to exp(1), corresponding to a log-scale value of 1.}

\item{input.vp23cmpt}{Initial estimate for second peripheral volume (Vp2).
Defaults to exp(1), corresponding to a log-scale value of 1.}

\item{input.q3cmpt}{Initial estimate for intercompartmental clearance between
central and first peripheral compartments (Q1). Defaults to exp(1),
corresponding to a log-scale value of 1.}

\item{input.q23cmpt}{Initial estimate for intercompartmental clearance between
central and second peripheral compartments (Q2). Defaults to exp(1),
corresponding to a log-scale value of 1.}

\item{input.add}{Additive error term. Defaults to 1.}
}
\value{
A list containing fitted parameter estimates and model prediction
error metrics.
}
\description{
Fits a three-compartment intravenous pharmacokinetic model using a naive
pooled data approach and evaluates model performance based on prediction
error metrics.
}
\details{
Rows with \code{EVID == 2} are excluded prior to model fitting. The model is
fitted using \code{Fit_3cmpt_iv}, and prediction-based metrics are calculated to
evaluate performance.
}
\examples{
\dontrun{
run_npd_3cmpt_iv(
  dat = Bolus_2CPT,
  input.cl = 4,
  input.vc3cmpt = 10,
  input.vp3cmpt = 10,
  input.vp23cmpt = 10,
  input.q3cmpt = 4,
  input.q23cmpt = 4
)
}

}
\seealso{
\code{\link{Fit_3cmpt_iv}}
}
