% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simpcal_iv.R
\name{run_simpcal_iv}
\alias{run_simpcal_iv}
\title{Run simplified calculation for intravenous pharmacokinetic data}
\usage{
run_simpcal_iv(dat, infusion_flag, sdflag, fdobsflag, half_life)
}
\arguments{
\item{dat}{A data frame containing the intravenous pharmacokinetic data.}

\item{infusion_flag}{A flag indicating whether the dataset includes infusion doses. Set to 1 if the data includes infusions, otherwise set to 0.}

\item{sdflag}{A flag indicating whether this is only single-dose data.}

\item{fdobsflag}{A flag indicating whether first dose observations are available.}

\item{half_life}{The half-life of the drug.}
}
\value{
A list containing the results of the simplified calculation, including clearance, volume of distribution, and the processed dataset.
}
\description{
Perform a simplified calculation for clearance and volume of distribution in intravenous cases.
}
\examples{
dat <- Bolus_1CPT
dat<- nmpkconvert(dat)
dat<- calculate_tad(dat)
run_simpcal_iv(dat, infusion_flag = 0, sdflag = 0, fdobsflag = 1)
}
