% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_simpcal_iv.R
\name{run_simpcal_iv}
\alias{run_simpcal_iv}
\title{Run approximate pharmacokinetic calculation for intravenous pharmacokinetic data}
\usage{
run_simpcal_iv(dat, route = "bolus", sdflag = 0, fdobsflag = 1, half_life = NA)
}
\arguments{
\item{dat}{A data frame containing the intravenous pharmacokinetic data.}

\item{route}{A character string specifying the route of administration (e.g, "bolus" and "infusion"). The default value is "bolus".}

\item{sdflag}{A flag indicating whether this is only single-dose data.}

\item{fdobsflag}{A flag indicating whether first dose observations are available.}

\item{half_life}{The half-life of the drug.}
}
\value{
A list containing the results of the simplified calculation, including clearance, volume of distribution, and the processed dataset.
}
\description{
Perform a simplified calculation for clearance and volume of distribution in intravenous cases.
}
\examples{

# Example 1 (iv case)
dat <- Bolus_1CPT
dat <- nmpkconvert(dat)
dat <- calculate_tad(dat)
run_simpcal_iv(dat)

# Example 2 (infusion case).

dat <- Infusion_1CPT
dat <- nmpkconvert(dat)
dat <- calculate_tad(dat,infusion_flag=1)
run_simpcal_iv(dat,route="infusion")
}
