% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_point.R
\name{run_single_point}
\alias{run_single_point}
\title{Run single-point pharmacokinetic analysis}
\usage{
run_single_point(dat, half_life = NA)
}
\arguments{
\item{dat}{A data frame containing pharmacokinetic data, including observed concentrations (\(DV\)), time after dose (\(tad\)), dosing interval (\(ii\)), and dose information.}

\item{half_life}{The half-life of the drug (\(t_{1/2}\)). If not provided, some calculations may use alternative assumptions or data-based rules.}
}
\value{
A list containing:
\itemize{
\item \code{singlepoint.results}: A data frame with estimated \(CL\), \(V_d\), \(k_a\), processing time, and a descriptive message.
\item \code{dat}: The input dataset used in the calculations.
\item \code{single_point_ka_df}: Data used for \(k_a\) calculations (if applicable).
\item \code{single_point_cl_df}: Data used for \(CL\) calculations.
\item \code{single_point_vd_df}: Data used for \(V_d\) calculations.
\item \code{approx.vc.out}: A message or result indicating whether the approximate central volume of distribution (\(V_d\)) was successfully estimated.
}
}
\description{
This function performs a single-point pharmacokinetic analysis, estimating key parameters such as clearance (\(CL\)), volume of distribution (\(V_d\)), and absorption rate constant (\(k_a\)) based on the provided dataset and optional half-life (\(t_{1/2}\)).
}
\details{
This function integrates two key components:
\enumerate{
\item \strong{\code{single_point_base}:} Conducts basic single-point calculations, determining parameters based on steady-state or dose-interval data.
\item \strong{\code{single_point_extra}:} Extends the analysis by handling additional cases, such as oral absorption rate (\(k_a\)) or compensating for missing \(CL\) or \(V_d\) values.
}

The function dynamically adapts to the provided data, leveraging different strategies based on the availability of steady-state, dose-interval, or single-dose data. If half-life (\(t_{1/2}\)) is provided, it is used in relevant calculations.
}
\examples{
# Example 1 (iv case)
dat <- Bolus_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)$half_life_median
run_single_point(fdat,half_life)

# Example 2 (infusion case).

dat <- Infusion_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)$half_life_median
single_point_base(fdat,half_life)
run_single_point(fdat,half_life)

dat <- Oral_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)$half_life_median
run_single_point(fdat,half_life)

}
