% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_point.R
\name{run_single_point}
\alias{run_single_point}
\title{Run single-point pharmacokinetic analysis}
\usage{
run_single_point(dat, half_life = NA)
}
\arguments{
\item{dat}{A data frame containing pharmacokinetic data, including observed concentrations (\code{DV}), time after dose (\code{tad}), dosing interval (\code{ii}), and dose information.}

\item{half_life}{The half-life of the drug (\eqn{t_{1/2}}). If not provided, alternative assumptions or data-based rules may be applied for calculations.}
}
\value{
A list containing:
\itemize{
\item \code{singlepoint.results}: A data frame with estimated \eqn{CL}, \eqn{V_d}, \eqn{k_a}, processing time, and a descriptive message.
\item \code{dat}: The input dataset used in the calculations.
\item \code{single_point_ka_df}: Data used for \eqn{k_a} calculations (if applicable).
\item \code{single_point_cl_df}: Data used for \eqn{CL} calculations.
\item \code{single_point_vd_df}: Data used for \eqn{V_d} calculations.
\item \code{approx.vc.out}: A message or result indicating whether the approximate central volume of distribution (\eqn{V_d}) was successfully estimated.
}
}
\description{
This function performs a single-point pharmacokinetic analysis, estimating key parameters such as clearance (\eqn{CL}), volume of distribution (\eqn{V_d}), and absorption rate constant (\eqn{k_a}) based on the provided dataset and optional half-life (\eqn{t_{1/2}}).
}
\details{
This function integrates two key components:
\itemize{
\item \code{single_point_base}: Conducts basic single-point calculations to estimate parameters based on steady-state or dose-interval data.
\item \code{single_point_extra}: Extends the analysis to handle additional cases, such as calculating oral absorption rate (\eqn{k_a}) or estimating parameters when data is incomplete (e.g., when \eqn{CL} or \eqn{V_d} is missing).
}
}
\examples{
# Example 1 (iv case)
dat <- Bolus_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)$half_life_median
run_single_point(fdat,half_life)

# Example 2 (infusion case).

dat <- Infusion_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)$half_life_median
run_single_point(fdat,half_life)

dat <- Oral_1CPT
fdat <- processData(dat)$dat
half_life<-half_life_estimated(dat = fdat)$half_life_median
run_single_point(fdat,half_life)

}
