% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sens_params.R
\name{sim_sens_2cmpt}
\alias{sim_sens_2cmpt}
\title{Sensitivity analysis for a two-compartment model}
\usage{
sim_sens_2cmpt(
  dat,
  sim_vc_list = NA,
  sim_vp_list = NA,
  estcl,
  estvd,
  sim_q_list = NA,
  estka = NA,
  noniv_flag = 0
)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data}

\item{sim_vc_list}{A list of central compartment volumes (Vc) to simulate (optional).}

\item{sim_vp_list}{A list of peripheral compartment volumes (Vp) to simulate (optional).}

\item{estcl}{Estimated clearance.}

\item{estvd}{Estimated volume of distribution (optional, required if \code{sim_vc_list} and \code{sim_vp_list} are not provided).}

\item{sim_q_list}{A list of inter-compartmental clearance (Q) to simulate (optional).}

\item{estka}{Estimated absorption rate if oral case.}
}
\value{
A data frame containing the Vc, Vp, APE, MAPE, and time spent for each simulation.
}
\description{
Performs sensitivity analysis by testing a series of potential ratios of vc to vp.
}
\examples{
dat <- Bolus_2CPT
sim.results<-sim_sens_2cmpt(dat, estka=1,estcl = 4, estvd = 70)
}
