% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sens_params.R
\name{sim_sens_2cmpt}
\alias{sim_sens_2cmpt}
\title{Sensitivity analysis for a two-compartment model}
\usage{
sim_sens_2cmpt(dat, sim_vc_list, sim_vp_list, estcl, estvd, estq)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data}

\item{sim_vc_list}{A list of central compartment volumes (Vc) to simulate (optional).}

\item{sim_vp_list}{A list of peripheral compartment volumes (Vp) to simulate (optional).}

\item{estcl}{Estimated clearance.}

\item{estvd}{Estimated volume of distribution (optional, required if \code{sim_vc_list} and \code{sim_vp_list} are not provided).}

\item{estq}{Estimated intercompartmental clearance (optional, default is \code{estcl}).}
}
\value{
A data frame containing the Vc, Vp, APE, MAPE, and time spent for each simulation.
}
\description{
This function performs sensitivity analysis by testing a series of potential ratios of vc to vp.
}
\examples{
dat <- Bolus_2CPT
sim_sens_2cmpt(dat, estcl = 4, estvd = 70)
}
