% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_sens_params.R
\name{sim_sens_3cmpt}
\alias{sim_sens_3cmpt}
\title{Sensitivity analysis for a three-compartment model}
\usage{
sim_sens_3cmpt(
  dat,
  sim_vc_list,
  sim_vp_list,
  sim_vp2_list,
  estcl,
  estvd,
  sim_q_list,
  estka = NA,
  noniv_flag = 0
)
}
\arguments{
\item{dat}{A data frame containing the pharmacokinetic data}

\item{sim_vc_list}{A list of central compartment volumes (Vc) to simulate (optional).}

\item{sim_vp_list}{A list of first peripheral compartment volumes (Vp) to simulate (optional).}

\item{sim_vp2_list}{A list of second peripheral compartment volumes (Vp2) to simulate (optional).}

\item{estcl}{Estimated clearance.}

\item{estvd}{Estimated volume of distribution (optional, required if \code{sim_vc_list}, \code{sim_vp_list}and \code{sim_vp2_list}  are not provided).}

\item{estq}{Estimated intercompartmental clearance (optional, default is \code{estcl}).}

\item{estq2}{Estimated intercompartmental clearance between central compartment and second peripheral compartment (optional, default is \code{estcl}).}
}
\value{
A data frame containing the Vc, Vp, Vp2 APE, MAPE, and time spent for each simulation.
}
\description{
This function performs sensitivity analysis by testing a series of potential ratios of vc to vp and vp to v2.
}
\examples{
dat <- Bolus_2CPT
sim_results<-sim_sens_3cmpt(dat, estka=1, estcl = 4, estvd = 70)
sim_results
}
