% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_single_point.R
\name{trimmed_geom_mean}
\alias{trimmed_geom_mean}
\title{Calculate the trimmed geometric mean}
\usage{
trimmed_geom_mean(x, trim = 0, na.rm = TRUE)
}
\arguments{
\item{x}{A numeric vector containing the values for which the trimmed geometric mean is to be calculated.}

\item{trim}{A numeric value between 0 and 0.5 indicating the proportion of observations
to be trimmed from each end of the vector. Default is 0 (no trimming).}

\item{na.rm}{Logical. Should missing values (NA) be removed before computation? Defaults to TRUE.}
}
\value{
The trimmed geometric mean of the input vector as a numeric value.
If no values remain after trimming, returns \code{NA}.
}
\description{
Computes the trimmed geometric mean of a numeric vector.
A specified proportion of the smallest and largest values is removed before calculating the geometric mean.
}
